+++ src/main/java/com/bsd/tutor/dao/CourseCalendarDao.java
@@ -32,7 +32,7 @@ public class CourseCalendarDao {
     }
 
 
-    public static Integer create(Class e) {
+    public static Long create(Class e) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
         session.save(e);
@@ -54,7 +54,7 @@ public class CourseCalendarDao {
     }
 
 
-    public static Class findByID(Integer id) {
+    public static Class findByID(Long id) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         Class e = (Class) session.load(Class.class, id);
         session.close();
